+++ misskey-api/CHANGELOG.md
@@ -63,6 +63,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for Misskey v12.104.0
    - endpoint `charts/ap-request`
 - Support for Misskey v12.105.0
+- Support for Misskey v12.106.0 ~ v12.106.3
 
 ### Changed
 ### Deprecated
@@ -81,6 +82,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    - For Misskey v12.102.0 ~
 - endpoint `charts/network`
    - For Misskey v12.104.0 ~
+- endpoint `stats`
+   - For Misskey v12.106.0 ~
 
 ### Fixed